restricted page if not login

+++ base/views.py
@@ -1,4 +1,5 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.db.models import Q
 from django.template import loader
 from django.http import HttpResponse
@@ -7,6 +8,8 @@ from .forms import RoomForm
 from django.contrib.auth.models import User
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+
 # Create your views here.
 
 # rooms = [
@@ -18,6 +21,9 @@ from django.contrib.auth import authenticate, login, logout
 
 def loginPage(request):
     context = {}
+
+    if request.user.is_authenticated:
+        return redirect('home')
 
     if request.method == 'POST':
         username = request.POST.get('username')
@@ -60,7 +66,7 @@ def home(request):
 def room(request):
     #template = loader.get_template("base/room.html")
     room = None
-    context = {'rooms': rooms};
+    context = {'room': room};
     return render(request, "base/room.html", context)
 
 def room_detail(request, pk):
@@ -71,7 +77,7 @@ def room_detail(request, pk):
 
     return render(request, "base/room_detail.html", context)
 
-
+@login_required(login_url='/login')
 def createRoom(request):
     form = RoomForm()
 
@@ -85,10 +91,13 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='/login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('Your are not allowed here!!')
 
     if request.method == 'POST':
         #print(request.POST)
@@ -101,9 +110,13 @@ def updateRoom(request, pk):
 
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='/login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
 
+    if request.user != room.host:
+        return HttpResponse('Your are not allowed here!!')
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
